External RHCS cluster Utilities for osd disruptive operations: 1. Reused most of the code from CephCI 2. Few utilities for kill and revive osds in an external RHCS cluster environment
